requirements.txt
torch>=1.7.0
torchvision>=0.8.1
numpy>=1.19.2
opencv-python>=4.4.0
pyyaml>=5.3.1
tqdm>=4.64.0
albumentations>=1.3.0